requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
netifaces==0.10.6
python-engineio==4.4.0
python-socketio==5.8.0
requests==2.28.2
urllib3==1.26.14
websocket-client==1.5.1
yarl==1.8.2
